backend/app/requirements.txt
flask
flask-cors
numpy
ortools
gunicorn
requests
locust
python-dotenv